requirements.txt
robotframework
pyyaml
nopea==0.0.4
